requirements.txt
Flask
gunicorn
google-cloud-secret-manager
google-cloud-storage
fsspec
gcsfs
python-gnupg
Werkzeug